More work on TaskBarNotifier. Files taskbarnotifier_dbus.cpp and taskbarnotifier_win.cpp merged to taskbarnotifier.cpp. Added USE_TASKBARNOTIFIER definition. Minor fixes.
