+++ src/widgets/CMakeLists.txt
@@ -63,13 +63,10 @@ set(HEADERS
     stretchwidget.h
 )
 
-if(LINUX AND USE_DBUS)
-    list(APPEND SOURCES taskbarnotifier_dbus.cpp)
+if(USE_DBUS OR WIN32)
+    list(APPEND SOURCES taskbarnotifier.cpp)
     list(APPEND HEADERS taskbarnotifier.h)
-elseif(WIN32)
-    list(APPEND SOURCES taskbarnotifier_win.cpp)
-    list(APPEND HEADERS taskbarnotifier.h)
-    if(${QT_DEFAULT_MAJOR_VERSION} VERSION_LESS "6")
+    if(WIN32 AND (${QT_DEFAULT_MAJOR_VERSION} VERSION_LESS "6"))
         find_package(Qt${QT_DEFAULT_MAJOR_VERSION} 5.10 REQUIRED COMPONENTS WinExtras)
         list(APPEND QT_LIBRARIES Qt${QT_DEFAULT_MAJOR_VERSION}::WinExtras)
     endif()
